requirements.txt
python-dxf==11.4.0
